fix(catalog): type homer's TSS peak_count as Int64, not polars' UInt64
`pl.len()` returns UInt64 and `EXPECTED_SCHEMA` declares Int64, and the recipe
schema check is exact, so `homer_tss_distance_profile` failed at ingest on both
chipseq and atacseq:

    Recipe homer/tss_distance_profile.py: column 'peak_count' expected Int64,
    got UInt64

The catalog fixture never caught it: it is a TSV, and the round-trip through
the file reads the column back as Int64.

Found by re-ingesting the three ChIP-family projects into the worktree stack;
all three now land every data collection (chipseq 15, atacseq 19, cutandrun 14).

+++ depictio/catalog/homer/tss_distance_profile.py
@@ -89,7 +89,10 @@ def transform(sources: dict[str, pl.DataFrame]) -> pl.DataFrame:
             ((pl.col("distance_to_tss") // BIN_BP) * BIN_BP + BIN_BP // 2).alias("distance_to_tss")
         )
         .group_by(["sample", "distance_to_tss"])
-        .agg(pl.len().alias("peak_count"))
+        # `pl.len()` is UInt64; EXPECTED_SCHEMA says Int64, and the schema check
+        # is exact, so cast here rather than letting the recipe fail at ingest
+        # (a TSV fixture round-trips back to Int64 and hides this).
+        .agg(pl.len().cast(pl.Int64).alias("peak_count"))
     )
     return (
         binned.with_columns(
